move content enhancers
