+++ src/components/header/ModuleHeader.tsx
@@ -1,4 +1,4 @@
-import { getModuleEmoji } from "@src/content/enhancers/moduleEmoji/getModuleEmoji"
+import { getModuleEmoji } from "@src/features/contentEnhancers/emoji/modules"
 import Title from "../typography/Title"
 import { extractUrlSegments } from "@src/features/versionControl"
