+++ src/features/contentEnhancers/emoji/files/index.ts
@@ -1,0 +1,3 @@
+export * from "./fileEmojiMap"
+export * from "./getFileEmoji"
+export * from "./getFileEmojiId"
