+++ src/features/contentEnhancers/emoji/modules/index.ts
@@ -1,0 +1,2 @@
+export * from "./getModuleEmoji"
+export * from "./moduleEmojiMap"
